+++ addon/components/object-list-view.js
@@ -407,6 +407,15 @@ export default FlexberryBaseComponent.extend(
   showEditMenuItemInRow: false,
 
   /**
+    Flag used to display filters in modal.
+
+    @property showFiltersInModal
+    @type Boolean
+    @default false
+  */
+  showFiltersInModal: false,
+
+  /**
     Flag indicates whether to show dropdown menu with prototype menu item, in last column of every row.
 
     @property showPrototypeMenuItemInRow
@@ -557,6 +566,7 @@ export default FlexberryBaseComponent.extend(
       ret = cols;
     }
 
+    this.get('objectlistviewEventsService').setOlvFilterColumnsArray(ret);
     return ret;
   }),