+++ addon/services/objectlistview-events.js
@@ -25,6 +25,15 @@ export default Ember.Service.extend(Ember.Evented, {
   _selectedRecords: undefined,
 
   /**
+    Current model projection columns with available filters.
+
+    @property _olvFilterColumnsArray
+    @type Object[]
+    @private
+  */
+  _olvFilterColumnsArray: undefined,
+
+  /**
     Init service.
 
     @method init
@@ -327,6 +336,26 @@ export default Ember.Service.extend(Ember.Evented, {
   */
   getLimitFunction(componentName) {
     return this.get(`currentLimitFunctions.${componentName}`);
+  },
+
+  /**
+    Sets olv columns with available filters.
+
+    @method setOlvFilterColumnsArray
+    @param Object[] Olv columns with available filters.
+  */
+  setOlvFilterColumnsArray(columns) {
+    this.set('_olvFilterColumnsArray', columns);
+  },
+
+  /**
+    Gets olv columns with available filters.
+
+    @method getOlvFilterColumnsArray
+    @param Object[] Olv columns with available filters.
+  */
+  getOlvFilterColumnsArray() {
+    return this.get('_olvFilterColumnsArray');
   },
 
   /**